pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-fetcher"
version = "0.1.0"
description = "CLI tool to batch-download files from a remote SFTP server"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "paramiko>=3.4",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "requests>=2.31",
    "google-genai>=0.2",
    "google-adk>=0.1",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pymysql>=1.1",
    "Pillow>=10.0",
    "watchdog>=4.0",
    "click>=8.1",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
file-fetcher = "file_fetcher.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
